+++ vite.config.ts
@@ -8,4 +8,14 @@ export default defineConfig({
 	resolve: {
 		alias: [{ find: '@', replacement: '/src' }],
 	},
+	// server: {
+	// 	https: true,
+	// 	proxy: {
+	// 		'api/': {
+	// 			target: 'https://api.dotoread.shop',
+	// 			secure: false,
+	// 			changeOrigin: true,
+	// 		},
+	// 	},
+	// },
 });
